// frontend/src/api/company/models/ApiCreateCompany.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApiCreateCompany = {
    /**
     * Название компании
     */
    name: string;
    /**
     * Описание компании
     */
    description?: string;
    /**
     * Адрес компании
     */
    address?: string;
    /**
     * Телефон компании
     */
    phone?: string;
    /**
     * Email компании
     */
    email?: string;
    /**
     * Сайт компании
     */
    website?: string;
    /**
     * Логотип компании
     */
    avatar?: string;
    /**
     * Уникальное название компании для короткой ссылки
     */
    short_link_name?: string;
};